pyproject.toml
[tool.poetry]
name = "nautobot-ssot-meraki"
version = "0.1.0"
description = "Nautobot SSoT for Meraki"
authors = ["Justin Drew <info@networktocode.com>"]
readme = "README.md"
homepage = "https://github.com/nautobot/nautobot-plugin-ssot-meraki"
repository = "https://github.com/nautobot/nautobot-plugin-ssot-meraki"
keywords = ["nautobot", "nautobot-plugin"]
include = [
    "README.md",
]
packages = [
    { include = "nautobot_ssot_meraki" },
]

[tool.poetry.dependencies]
python = "^3.8, <3.12"
# Used for local development
nautobot = { version = "^1.6.0", optional = true }
meraki = "^1.37.2"
nautobot-ssot = "^1.5.0"

[tool.poetry.dev-dependencies]
bandit = "*"
black = "*"
coverage = "*"
django-debug-toolbar = "*"
# we need to pin flake8 because of package dependencies that cause it to downgrade and
# therefore cause issues with linting since older versions do not take .flake8 as config
flake8 = "^3.9.2"
invoke = "*"
ipython = "*"
pydocstyle = "*"
pylint = "*"
pylint-django = "*"
pylint-nautobot = "*"
pytest = "*"
python-dotenv = "^0.21.1"
yamllint = "*"
Markdown = "*"
toml = "*"
# Rendering docs to HTML
mkdocs = "1.4.3"
# Material for MkDocs theme
mkdocs-material = "9.1.15"
# Render custom markdown for version added/changed/remove notes
mkdocs-version-annotations = "1.0.0"
python-semantic-release = "^7.34.3"

[tool.poetry.extras]
nautobot = ["nautobot"]

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django, pylint_nautobot"
ignore=".venv"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[tool.pylint-nautobot]
supported_nautobot_versions = [
    "1.6.0"
]

[tool.pydocstyle]
convention = "google"
inherit = false
match = "(?!__init__).*\\.py"
match-dir = "(?!tests|migrations|development)[^\\.].*"
# D212 is enabled by default in google convention, and complains if we have a docstring like:
# """
# My docstring is on the line after the opening quotes instead of on the same line as them.
# """
# We've discussed and concluded that we consider this to be a valid style choice.
add_ignore = "D212"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = "-vv --doctest-modules"

[tool.semantic_release]
version_toml = [
    "pyproject.toml:tool.poetry.version"
]
major_on_zero = true
branch = "main"
upload_to_repository = false
upload_to_release = false
changelog_file = "CHANGELOG.md"
build_command = false
